move all functions to util
